Moved AddButton component to the Button folder
